Potential fix for code scanning alert no. 151: Multiplication result converted to larger type
multiplication result fixed via autofix

+++ crypto/vm/boc.cpp
@@ -427,7 +427,7 @@ std::size_t BagOfCells::estimate_serialized_size(int mode) {
   info.absent_count = dangle_count;
   int crc_size = info.has_crc32c ? 4 : 0;
   info.roots_offset = 4 + 1 + 1 + 3 * info.ref_byte_size + info.offset_byte_size;
-  info.index_offset = info.roots_offset + info.root_count * info.ref_byte_size;
+  info.index_offset = info.roots_offset + static_cast<long long>(info.root_count) * info.ref_byte_size;
   info.data_offset = info.index_offset;
   if (info.has_index) {
     info.data_offset += (long long)cell_count * info.offset_byte_size;
